+++ extensions/bunch/src/types/preferences.ts
@@ -1,4 +1,3 @@
 export interface Preferences {
   rememberFilter: boolean;
-  closeMainWindow: boolean;
 }